refactor(player): extract Session + Achievement managers from GameDetailViewModel (#691)
`GameDetailViewModel.kt` was 894 LOC mixing 8+ unrelated domains:
metadata, sessions CRUD, achievements (3 view modes), cheats, saves,
challenges, ratings, shared saves, collections.

Extracted two self-contained slices into their own manager classes,
following the delegation pattern already used by EmulationViewModel
(SaveManager / ChallengeManager / NetplayManager):

1. **GameDetailSessionManager** (120 LOC) — loadSessions,
   createSession, renameSession, deleteSession, cloneSession.
   5 methods, all session-scoped CRUD.

2. **GameDetailAchievementManager** (87 LOC) — loadAchievements,
   loadAchievementTimeline, loadAchievementLeaderboard,
   toggleAchievementsView. The "cache-once-then-reuse" logic for
   each of the three view modes.

Both managers share the VM's `_state` StateFlow so updates land
directly without a merge step. The VM delegates each intent through
the managers in `onIntent` and calls them from internal code
paths too (e.g. `loadGame` seeds achievements/sessions,
`adminRefreshAchievements` reloads via the manager).

LOC:
- GameDetailViewModel.kt: 894 → 765 (-129)
- New: GameDetailSessionManager.kt (120), GameDetailAchievementManager.kt (87)
- Net: +78 LOC of manager scaffolding, but the VM's remaining 765
  lines are now more focused on metadata / downloads / ratings /
  shared saves / collections.

Behaviour preserved. 526 shared + 759 desktop tests pass. Both
targets compile (desktop + Android). detektMainDesktop clean.

Deferred (mentioned in #691): shared saves (~70 LOC) + ratings
(~40 LOC) clusters are smaller; worth a `SocialManager` only if
more social features land. Monitor.

Closes #691
